Moved UserSearchResult component to u.tsx
